requirements.txt
# Core
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.19
pillow>=8.0
packaging>=20.0

# Visualization / HTML
visdom>=0.2.4
dominate>=2.6.0

# Data / Tools
opencv-python>=4.5
requests>=2.25
beautifulsoup4>=4.9
lxml>=4.6
scipy>=1.5
scikit-learn>=0.24
tifffile>=2021.7.2
GPUtil>=1.4.0
tqdm>=4.62

# Evaluation (optional)
pytorch-fid>=0.3.0
